Move macros into separate module
